requirements.dev.txt
# Code Quality
flake8>=3.9.0,<4.0
flake8-django>=1.4.0,<1.5
flake8-import-order>=0.18.2,<0.19
flake8-quotes>=3.4.0,<3.5

# Django
django-debug-toolbar>=3.0,<4.0
# django-extensions>=3.0,<4.0
# django-cors-headers>=3.0,<4.0
requests>=2.28.1,<3.0
beautifulsoup4>=4.12.2,<5.0
